code-server/requirements.txt
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
click==8.2.1
exceptiongroup==1.3.0
fastapi==0.115.12
h11==0.16.0
idna==3.10
mpmath==1.3.0
numpy==2.2.6
omegaconf==2.3.0
packaging==25.0
pydantic==2.11.5
pydantic-core==2.33.2
python-dotenv==1.1.0
pyyaml==6.0.2
restrictedpython==8.0
shortuuid==1.0.13
sniffio==1.3.1
starlette==0.46.2
sympy==1.14.0
typing-extensions==4.13.2
typing-inspection==0.4.1
uvicorn==0.34.2
